缓慢修改屎山代码

+++ cpp/Logic/Army.cpp
@@ -9,7 +9,7 @@ Army::Army(int force, int level) {
 Army::~Army() {
 }
 
-float Army::getSpeed(){
+double Army::getSpeed() {
 	return this->speed;
 }
