+++ cpp/Logic/Weapon.cpp
@@ -9,10 +9,10 @@ Weapon::Weapon(int id) {
 	Config& config = Config::getInstance();
 	std::string ID = std::to_string(id);
 	const json j = config.getConfig({ "Weapon", ID});
-	this->damage = j["damage"].get<float>();
-	this->damageRange = j["damageRange"].get<float>();
-	this->attackSpeed = j["attackSpeed"].get<float>();
-	this->attackRange = std::make_tuple(j["AttackRange"][0].get<float>(), j["AttackRange"][1].get<float>());
+	this->damage = j["damage"].get<double>();
+	this->damageRange = j["damageRange"].get<double>();
+	this->attackSpeed = j["attackSpeed"].get<double>();
+	this->attackRange = std::make_tuple(j["AttackRange"][0].get<double>(), j["AttackRange"][1].get<double>());
 	this->cost = j["cost"].get<std::vector<int>>();
 	this->id = id;
 	this->AICost = j["AICost"].get<int>();
@@ -21,23 +21,23 @@ Weapon::Weapon(int id) {
 Weapon::~Weapon() {
 }
 
-float Weapon::getDamage(int level){
+double Weapon::getDamage(int level) {
 	//formula
-	float Damage = this->damage * (1 + level * 0.2);
+	double Damage = this->damage * (1 + level * 0.2);
 	return Damage;
 }
 
-float Weapon::getDamageRange(int level){
-	float DamageRange = this->damageRange * (1 + level * 0.2);	
+double Weapon::getDamageRange(int level) {
+	double DamageRange = this->damageRange * (1 + level * 0.2);
 	return DamageRange;
 }
 
-float Weapon::getAttackSpeed(int level){
-	float AttackSpeed = this->attackSpeed * (1 + level * 0.2) * RegionManager::getInstance().get_map_width();
+double Weapon::getAttackSpeed(int level) {
+	double AttackSpeed = this->attackSpeed * (1 + level * 0.2) * RegionManager::getInstance().get_map_width();
 	return AttackSpeed; 
 }
 
-std::tuple<float, float> Weapon::getAttackRange(){
+std::tuple<double, double> Weapon::getAttackRange() {
 	return this->attackRange;
 }
